+++ src/auth/index.js
@@ -1,11 +1,9 @@
 export {
-  isUserSignedIn, redirectToSignIn, redirectToSignInWithAuthRequest,
-  getAuthResponseToken, isSignInPending,
-  handlePendingSignIn, loadUserData, signUserOut,
-  generateAndStoreTransitKey, getTransitKey
-} from './authApp'
+  AppConfig
+} from './appConfig'
+
 export {
-  makeAuthRequest, makeAuthResponse
+  makeAuthResponse
 } from './authMessages'
 export {
   getAuthRequestFromURL, fetchAppManifest, redirectUserToApp