+++ src/network.js
@@ -148,7 +148,9 @@ export class BlockstackNetwork {
   }
 
   getNameInfo(fullyQualifiedName: string) {
-    return fetch(`${this.blockstackAPIUrl}/v1/names/${fullyQualifiedName}`)
+    console.log(this.blockstackAPIUrl)
+    const nameLookupURL = `${this.blockstackAPIUrl}/v1/names/${fullyQualifiedName}`
+    return fetch(nameLookupURL)
       .then((resp) => {
         if (resp.status === 404) {
           throw new Error('Name not found')
@@ -159,6 +161,7 @@ export class BlockstackNetwork {
         }
       })
       .then((nameInfo) => {
+        console.log(`nameInfo: ${JSON.stringify(nameInfo)}`)
         // the returned address _should_ be in the correct network ---
         //  blockstackd gets into trouble because it tries to coerce back to mainnet
         //  and the regtest transaction generation libraries want to use testnet addresses