+++ src/storage/index.js
@@ -10,12 +10,15 @@ import {
 import {
   encryptECIES, decryptECIES, signECDSA, verifyECDSA
 } from '../encryption'
-import { loadUserData } from '../auth'
 import { getPublicKeyFromPrivate, publicKeyToAddress } from '../keys'
 import { lookupProfile } from '../profiles'
-
-import { SignatureVerificationError } from '../errors'
+import {
+  InvalidStateError,
+  SignatureVerificationError
+} from '../errors'
 import { Logger } from '../logger'
+
+import type { UserSession } from '../auth/userSession'
 
 const SIGNATURE_FILE_SUFFIX = '.sig'
 
@@ -56,17 +59,22 @@ export function getUserAppFileUrl(path: string, username: string, appOrigin: str
 
 /**
  * Encrypts the data provided with the app public key.
+ * @param {UserSession} caller - the instance calling this method
  * @param {String|Buffer} content - data to encrypt
  * @param {Object} [options=null] - options object
  * @param {String} options.publicKey - the hex string of the ECDSA public
  * key to use for encryption. If not provided, will use user's appPrivateKey.
  * @return {String} Stringified ciphertext object
+ * @private
  */
-export function encryptContent(content: string | Buffer, options?: {publicKey?: string}) {
+export function encryptContentImpl(caller: UserSession,
+                                   content: string | Buffer,
+                                   options?: {publicKey?: string}) {
   const defaults = { publicKey: null }
   const opt = Object.assign({}, defaults, options)
   if (!opt.publicKey) {
-    const privateKey = loadUserData().appPrivateKey
+    const userData = caller.loadUserData()
+    const privateKey = userData.appPrivateKey
     opt.publicKey = getPublicKeyFromPrivate(privateKey)
   }
 
@@ -77,18 +85,22 @@ export function encryptContent(content: string | Buffer, options?: {publicKey?: 
 /**
  * Decrypts data encrypted with `encryptContent` with the
  * transit private key.
+ * @param {UserSession} caller - the instance calling this method
  * @param {String|Buffer} content - encrypted content.
  * @param {Object} [options=null] - options object
  * @param {String} options.privateKey - the hex string of the ECDSA private
  * key to use for decryption. If not provided, will use user's appPrivateKey.
  * @return {String|Buffer} decrypted content.
+ * @private
  */
-export function decryptContent(content: string, options?: {privateKey?: ?string}) {
+export function decryptContentImpl(caller: UserSession,
+                                   content: string,
+                                   options?: {privateKey?: ?string}) {
   const defaults = { privateKey: null }
   const opt = Object.assign({}, defaults, options)
   let privateKey = opt.privateKey
   if (!privateKey) {
-    privateKey = loadUserData().appPrivateKey
+    privateKey = caller.loadUserData().appPrivateKey
   }
 
   try {
@@ -108,13 +120,14 @@ export function decryptContent(content: string, options?: {privateKey?: ?string}
  * (username, app) pair.
  * @private
  */
-function getGaiaAddress(app: string, username: ?string, zoneFileLookupURL: ?string) {
+function getGaiaAddress(caller: UserSession,
+                        app: string, username: ?string, zoneFileLookupURL: ?string) {
   return Promise.resolve()
     .then(() => {
       if (username) {
         return getUserAppFileUrl('/', username, app, zoneFileLookupURL)
       } else {
-        return getOrSetLocalGaiaHubConnection()
+        return getOrSetLocalGaiaHubConnection(caller)
           .then(gaiaHubConfig => getFullReadUrl('/', gaiaHubConfig))
       }
     })
@@ -131,14 +144,15 @@ function getGaiaAddress(app: string, username: ?string, zoneFileLookupURL: ?stri
  *  multi-player reads and reads from own storage.
  * @private
  */
-function getFileContents(path: string, app: string, username: ?string, zoneFileLookupURL: ?string,
+function getFileContents(caller: UserSession,
+                         path: string, app: string, username: ?string, zoneFileLookupURL: ?string,
                          forceText: boolean) : Promise<?string | ?ArrayBuffer> {
   return Promise.resolve()
     .then(() => {
       if (username) {
         return getUserAppFileUrl(path, username, app, zoneFileLookupURL)
       } else {
-        return getOrSetLocalGaiaHubConnection()
+        return getOrSetLocalGaiaHubConnection(caller)
           .then(gaiaHubConfig => getFullReadUrl(path, gaiaHubConfig))
       }
     })
@@ -175,7 +189,7 @@ function getFileContents(path: string, app: string, username: ?string, zoneFileL
  *  from own storage.
  * @private
  */
-function getFileSignedUnencrypted(path: string, opt: {
+function getFileSignedUnencrypted(caller: UserSession, path: string, opt: {
   username?: ?string,
   app: string,
   zoneFileLookupURL?: ?string
@@ -185,10 +199,10 @@ function getFileSignedUnencrypted(path: string, opt: {
   //    profile lookups to figure out where to read files
   //    do browsers cache all these requests if Content-Cache is set?
   return Promise.all(
-    [getFileContents(path, opt.app, opt.username, opt.zoneFileLookupURL, false),
-     getFileContents(`${path}${SIGNATURE_FILE_SUFFIX}`, opt.app, opt.username,
+    [getFileContents(caller, path, opt.app, opt.username, opt.zoneFileLookupURL, false),
+     getFileContents(caller, `${path}${SIGNATURE_FILE_SUFFIX}`, opt.app, opt.username,
                      opt.zoneFileLookupURL, true),
-     getGaiaAddress(opt.app, opt.username, opt.zoneFileLookupURL)]
+     getGaiaAddress(caller, opt.app, opt.username, opt.zoneFileLookupURL)]
   )
     .then(([fileContents, signatureContents, gaiaAddress]) => {
       if (!fileContents) {
@@ -239,14 +253,14 @@ function getFileSignedUnencrypted(path: string, opt: {
  *  gaia address for verification of the claimed public key.
  * @private
  */
-function handleSignedEncryptedContents(path: string, storedContents: string,
+function handleSignedEncryptedContents(caller: UserSession, path: string, storedContents: string,
                                        app: string, username: ?string, zoneFileLookupURL: ?string) {
-  const appPrivateKey = loadUserData().appPrivateKey
+  const appPrivateKey = caller.loadUserData().appPrivateKey
   const appPublicKey = getPublicKeyFromPrivate(appPrivateKey)
 
   let addressPromise
   if (username) {
-    addressPromise = getGaiaAddress(app, username, zoneFileLookupURL)
+    addressPromise = getGaiaAddress(caller, app, username, zoneFileLookupURL)
   } else {
     const address = publicKeyToAddress(appPublicKey)
     addressPromise = Promise.resolve(address)
@@ -286,13 +300,14 @@ function handleSignedEncryptedContents(path: string, storedContents: string,
       throw new SignatureVerificationError('Contents do not match ECDSA signature in file:'
                                            + ` ${path}`)
     } else {
-      return decryptContent(cipherText)
+      return caller.decryptContent(cipherText)
     }
   })
 }
 
 /**
  * Retrieves the specified file from the app's data store.
+ * @param {UserSession} caller - instance calling this method
  * @param {String} path - the path to the file to read
  * @param {Object} [options=null] - options object
  * @param {Boolean} [options.decrypt=true] - try to decrypt the data with the app private key
@@ -306,19 +321,24 @@ function handleSignedEncryptedContents(path: string, storedContents: string,
  * blockstack.js's getNameInfo function instead.
  * @returns {Promise} that resolves to the raw data in the file
  * or rejects with an error
+ * @private
  */
-export function getFile(path: string, options?: {
+export function getFileImpl(caller: UserSession, path: string, options?: {
     decrypt?: boolean,
     verify?: boolean,
     username?: string,
     app?: string,
     zoneFileLookupURL?: ?string
   }) {
+  const appConfig = caller.session.appConfig
+  if (!appConfig) {
+    throw new InvalidStateError('Missing AppConfig')
+  }
   const defaults = {
     decrypt: true,
     verify: false,
     username: null,
-    app: window.location.origin,
+    app: appConfig.appDomain,
     zoneFileLookupURL: null
   }
 
@@ -327,10 +347,10 @@ export function getFile(path: string, options?: {
   // in the case of signature verification, but no
   //  encryption expected, need to fetch _two_ files.
   if (opt.verify && !opt.decrypt) {
-    return getFileSignedUnencrypted(path, opt)
+    return getFileSignedUnencrypted(caller, path, opt)
   }
 
-  return getFileContents(path, opt.app, opt.username, opt.zoneFileLookupURL, !!opt.decrypt)
+  return getFileContents(caller, path, opt.app, opt.username, opt.zoneFileLookupURL, !!opt.decrypt)
     .then((storedContents) => {
       if (storedContents === null) {
         return storedContents
@@ -338,12 +358,12 @@ export function getFile(path: string, options?: {
         if (typeof storedContents !== 'string') {
           throw new Error('Expected to get back a string for the cipherText')
         }
-        return decryptContent(storedContents)
+        return decryptContentImpl(caller, storedContents)
       } else if (opt.decrypt && opt.verify) {
         if (typeof storedContents !== 'string') {
           throw new Error('Expected to get back a string for the cipherText')
         }
-        return handleSignedEncryptedContents(path, storedContents,
+        return handleSignedEncryptedContents(caller, path, storedContents,
                                              opt.app, opt.username, opt.zoneFileLookupURL)
       } else if (!opt.verify && !opt.decrypt) {
         return storedContents
@@ -355,6 +375,7 @@ export function getFile(path: string, options?: {
 
 /**
  * Stores the data provided in the app's data store to to the file specified.
+ * @param {UserSession} caller - instance calling this method
  * @param {String} path - the path to store the data in
  * @param {String|Buffer} content - the data to store in the file
  * @param {Object} [options=null] - options object
@@ -364,8 +385,9 @@ export function getFile(path: string, options?: {
  *                                         the app private key
  * @return {Promise} that resolves if the operation succeed and rejects
  * if it failed
+ * @private
  */
-export function putFile(path: string, content: string | Buffer, options?: {
+export function putFileImpl(caller: UserSession, path: string, content: string | Buffer, options?: {
   encrypt?: boolean | string,
   sign?: boolean
   }) {
@@ -390,7 +412,7 @@ export function putFile(path: string, content: string | Buffer, options?: {
     if (typeof (opt.sign) === 'string') {
       privateKey = opt.sign
     } else {
-      privateKey = loadUserData().appPrivateKey
+      privateKey = caller.loadUserData().appPrivateKey
     }
   }
   if (opt.encrypt) {
@@ -398,7 +420,7 @@ export function putFile(path: string, content: string | Buffer, options?: {
       publicKey = opt.encrypt
     } else {
       if (!privateKey) {
-        privateKey = loadUserData().appPrivateKey
+        privateKey = caller.loadUserData().appPrivateKey
       }
       publicKey = getPublicKeyFromPrivate(privateKey)
     }
@@ -410,7 +432,7 @@ export function putFile(path: string, content: string | Buffer, options?: {
   if (!opt.encrypt && opt.sign) {
     const signatureObject = signECDSA(privateKey, content)
     const signatureContent = JSON.stringify(signatureObject)
-    return getOrSetLocalGaiaHubConnection()
+    return getOrSetLocalGaiaHubConnection(caller)
       .then(gaiaHubConfig => Promise.all([
         uploadToGaiaHub(path, content, gaiaHubConfig, contentType),
         uploadToGaiaHub(`${path}${SIGNATURE_FILE_SUFFIX}`,
@@ -420,10 +442,10 @@ export function putFile(path: string, content: string | Buffer, options?: {
 
   // In all other cases, we only need one upload.
   if (opt.encrypt && !opt.sign) {
-    content = encryptContent(content, { publicKey })
+    content = encryptContentImpl(caller, content, { publicKey })
     contentType = 'application/json'
   } else if (opt.encrypt && opt.sign) {
-    const cipherText = encryptContent(content, { publicKey })
+    const cipherText = encryptContentImpl(caller, content, { publicKey })
     const signatureObject = signECDSA(privateKey, cipherText)
     const signedCipherObject = {
       signature: signatureObject.signature,
@@ -433,7 +455,7 @@ export function putFile(path: string, content: string | Buffer, options?: {
     content = JSON.stringify(signedCipherObject)
     contentType = 'application/json'
   }
-  return getOrSetLocalGaiaHubConnection()
+  return getOrSetLocalGaiaHubConnection(caller)
     .then(gaiaHubConfig => uploadToGaiaHub(path, content, gaiaHubConfig, contentType))
 }
 
@@ -446,17 +468,6 @@ export function putFile(path: string, content: string | Buffer, options?: {
  */
 export function getAppBucketUrl(gaiaHubUrl: string, appPrivateKey: string) {
   return getBucketUrl(gaiaHubUrl, appPrivateKey)
-}
-
-/**
- * Deletes the specified file from the app's data store. Currently not implemented.
- * @param {String} path - the path to the file to delete
- * @returns {Promise} that resolves when the file has been removed
- * or rejects with an error
- * @private
- */
-export function deleteFile(path: string) {
-  Promise.reject(new Error(`Delete of ${path} not supported by gaia hubs`))
 }
 
 /**
@@ -533,12 +544,15 @@ function listFilesLoop(hubConfig: GaiaHubConfig,
 
 /**
  * List the set of files in this application's Gaia storage bucket.
+ * @param {UserSession} caller - instance calling this method
  * @param {function} callback - a callback to invoke on each named file that
  * returns `true` to continue the listing operation or `false` to end it
  * @return {Promise} that resolves to the number of files listed
+ * @private
  */
-export function listFiles(callback: (name: string) => boolean) : Promise<number> {
-  return getOrSetLocalGaiaHubConnection()
+export function listFilesImpl(caller: UserSession,
+                              callback: (name: string) => boolean) : Promise<number> {
+  return getOrSetLocalGaiaHubConnection(caller)
     .then(gaiaHubConfig => listFilesLoop(gaiaHubConfig, null, 0, 0, callback))
 }