+++ tests/unitTests/src/unitTestsAuth.js
@@ -6,7 +6,6 @@ import FetchMock from 'fetch-mock'
 import {
   makeECPrivateKey,
   getPublicKeyFromPrivate,
-  makeAuthRequest,
   makeAuthResponse,
   verifyAuthRequest,
   verifyAuthResponse,
@@ -20,11 +19,9 @@ import {
   isManifestUriValid,
   isRedirectUriValid,
   verifyAuthRequestAndLoadManifest,
-  handlePendingSignIn,
-  signUserOut
+  UserSession,
+  AppConfig
 } from '../../../lib'
-
-import { BLOCKSTACK_APP_PRIVATE_KEY_LABEL } from '../../../lib/auth/authConstants'
 
 import { sampleProfiles, sampleNameRecords } from './sampleData'
 
@@ -33,18 +30,15 @@ global.window = {}
 export function runAuthTests() {
   const privateKey = 'a5c61c6ca7b3e7e55edee68566aeab22e4da26baa285c7bd10e8d2218aa3b229'
   const publicKey = '027d28f9951ce46538951e3697c62588a87f1f1f295de4a14fdd4c780fc52cfe69'
-  const nameLookupURL = 'https://explorer-api.appartisan.com/get_name_blockchain_record/'
+  const nameLookupURL = 'https://core.blockstack.org/v1/names/'
 
   test('makeAuthRequest && verifyAuthRequest', (t) => {
     t.plan(15)
 
-    global.window.location = {
-      origin: 'http://localhost:3000',
-      hostname: 'localhost',
-      host: 'localhost:3000',
-      href: 'http://localhost:3000/signin'
-    }
-    const authRequest = makeAuthRequest(privateKey)
+    const appConfig = new AppConfig('http://localhost:3000')
+    const blockstack = new UserSession(appConfig)
+
+    const authRequest = blockstack.makeAuthRequest(privateKey)
     t.ok(authRequest, 'auth request should have been created')
     console.log(authRequest)
 
@@ -60,7 +54,7 @@ export function runAuthTests() {
     t.equal(decodedToken.payload.domain_name,
             origin, 'auth request domain_name should be origin')
     t.equal(decodedToken.payload.redirect_uri,
-            'http://localhost:3000/', 'auth request redirects to correct uri')
+            'http://localhost:3000', 'auth request redirects to correct uri')
     t.equal(decodedToken.payload.manifest_uri,
             'http://localhost:3000/manifest.json', 'auth request manifest is correct uri')
 
@@ -104,7 +98,10 @@ export function runAuthTests() {
   test('invalid auth request - signature not verified', (t) => {
     t.plan(3)
 
-    const authRequest = makeAuthRequest(privateKey, 'http://localhost:3000')
+    const appConfig = new AppConfig('http://localhost:3000')
+    const blockstack = new UserSession(appConfig)
+
+    const authRequest = blockstack.makeAuthRequest(privateKey)
     const invalidAuthRequest = authRequest.substring(0, authRequest.length - 1)
 
     t.equal(doSignaturesMatchPublicKeys(invalidAuthRequest), false,
@@ -126,9 +123,12 @@ export function runAuthTests() {
 
   test('invalid auth request - invalid redirect uri', (t) => {
     t.plan(3)
+    const appConfig = new AppConfig('http://localhost:3000')
+    appConfig.redirectURI = () => 'https://example.com' // monkey patch for test
+    const blockstack = new UserSession(appConfig)
 
-    const invalidAuthRequest = makeAuthRequest(privateKey, 'https://example.com')
-
+    const invalidAuthRequest = blockstack.makeAuthRequest(privateKey)
+    console.log(invalidAuthRequest)
     t.equal(isRedirectUriValid(invalidAuthRequest), false,
             'Redirect URI should be invalid since it does not match origin')
 
@@ -149,7 +149,10 @@ export function runAuthTests() {
   test('invalid auth request - invalid manifest uri', (t) => {
     t.plan(2)
 
-    const invalidAuthRequest = makeAuthRequest(privateKey, 'http://localhost:3000', 'https://example.com/manifest.json')
+    const appConfig = new AppConfig('http://localhost:3000')
+    appConfig.manifestURI = () => 'https://example.com/manifest.json' // monkey patch for test
+    const blockstack = new UserSession(appConfig)
+    const invalidAuthRequest = blockstack.makeAuthRequest(privateKey)
 
     t.equal(isManifestUriValid(invalidAuthRequest), false,
             'Manifest URI should be invalid since it does not match origin')
@@ -224,6 +227,9 @@ export function runAuthTests() {
   test('auth response with invalid private key', (t) => {
     t.plan(2)
 
+    const appConfig = new AppConfig('http://localhost:3000')
+    const blockstack = new UserSession(appConfig)
+
     const url = `${nameLookupURL}ryan.id`
     // console.log(`URL: ${url}`)
 
@@ -233,24 +239,15 @@ export function runAuthTests() {
     const transitPrivateKey = makeECPrivateKey()
     const transitPublicKey = getPublicKeyFromPrivate(transitPrivateKey)
     const badTransitPrivateKey = makeECPrivateKey()
+    blockstack.session.transitKey = badTransitPrivateKey
     const metadata = { }
 
     const authResponse = makeAuthResponse(privateKey, sampleProfiles.ryan, 'ryan.id',
                                           metadata, undefined, appPrivateKey, undefined,
                                           transitPublicKey)
-    // console.log(decodeToken(authResponse))
-    global.window = Object.assign({ }, global.window, {
-      location: {
-        search: `authResponse=${authResponse}`
-      }
-    })
 
-    global.location = global.window.location
 
-    global.localStorage.setItem(BLOCKSTACK_APP_PRIVATE_KEY_LABEL,
-                                badTransitPrivateKey)
-
-    handlePendingSignIn(nameLookupURL)
+    blockstack.handlePendingSignIn(authResponse)
       .then(() => {
         t.fail('Should have failed to decrypt auth response')
       })
@@ -259,9 +256,9 @@ export function runAuthTests() {
         t.pass('Should fail to decrypt auth response')
       })
       .then(() => {
-        global.window.localStorage.setItem(BLOCKSTACK_APP_PRIVATE_KEY_LABEL,
-                                           transitPrivateKey)
-        return handlePendingSignIn(nameLookupURL)
+        blockstack.session.transitKey = transitPrivateKey
+
+        return blockstack.handlePendingSignIn(authResponse)
       })
       .then(() => {
         t.pass('Should correctly sign in with correct transit key')
@@ -270,25 +267,6 @@ export function runAuthTests() {
         console.log(err.stack)
         t.fail('Should not error')
       })
-  })
-
-  test('signUserOut with redirect', (t) => {
-    t.plan(1)
-    const startURL = 'https://example.com'
-    const redirectURL = 'https://example.com/redirect'
-    window.location = startURL
-
-    signUserOut(redirectURL)
-    t.equal(redirectURL, window.location, 'User should be redirected to the redirectURL')
-  })
-
-  test('signUserOut without redirect', (t) => {
-    t.plan(1)
-    const startURL = 'https://example.com'
-    window.location = startURL
-
-    signUserOut()
-    t.equal(startURL, window.location, 'User should not be redirected')
   })
 
   test('handlePendingSignIn with authResponseToken', (t) => {
@@ -303,12 +281,15 @@ export function runAuthTests() {
     const transitPublicKey = getPublicKeyFromPrivate(transitPrivateKey)
     const metadata = {}
 
+    const appConfig = new AppConfig('http://localhost:3000')
+    const blockstack = new UserSession(appConfig)
+    blockstack.session.transitKey = transitPrivateKey
+
     const authResponse = makeAuthResponse(privateKey, sampleProfiles.ryan, 'ryan.id',
                                           metadata, undefined, appPrivateKey, undefined,
                                           transitPublicKey)
-    global.window.localStorage.setItem(BLOCKSTACK_APP_PRIVATE_KEY_LABEL,
-                                       transitPrivateKey)
-    handlePendingSignIn(nameLookupURL, authResponse)
+
+    blockstack.handlePendingSignIn(authResponse)
       .then(() => {
         t.pass('Should correctly sign in with auth response')
       })
@@ -318,7 +299,7 @@ export function runAuthTests() {
       })
   })
 
-  test('handlePendingSignIn with authResponseToken and transit key', (t) => {
+  test('handlePendingSignIn 2', (t) => {
     t.plan(1)
 
     const url = `${nameLookupURL}ryan.id`
@@ -334,7 +315,11 @@ export function runAuthTests() {
                                           metadata, undefined, appPrivateKey, undefined,
                                           transitPublicKey)
 
-    handlePendingSignIn(nameLookupURL, authResponse, transitPrivateKey)
+    const appConfig = new AppConfig('http://localhost:3000')
+    const blockstack = new UserSession(appConfig)
+    blockstack.session.transitKey = transitPrivateKey
+
+    blockstack.handlePendingSignIn(authResponse)
       .then(() => {
         t.pass('Should correctly sign in with auth response')
       })
@@ -342,5 +327,39 @@ export function runAuthTests() {
         console.log(err.stack)
         t.fail('Should not error')
       })
+  })
+
+  test('app config defaults app domain to origin', (t) => {
+    t.plan(5)
+    global.window = {
+      location: {
+        origin: 'https://example.com'
+      }
+    }
+
+    const appConfig = new AppConfig()
+
+    t.equal(appConfig.appDomain, 'https://example.com')
+    t.equal(appConfig.scopes.length, 1)
+    t.equal(appConfig.scopes[0], 'store_write')
+    t.equal(appConfig.manifestURI(), 'https://example.com/manifest.json')
+    t.equal(appConfig.redirectURI(), 'https://example.com')
+  })
+
+  test('app config works with custom app domain to origin', (t) => {
+    t.plan(5)
+    global.window = {
+      location: {
+        origin: 'https://example.com'
+      }
+    }
+
+    const appConfig = new AppConfig('https://custom.example.com')
+
+    t.equal(appConfig.appDomain, 'https://custom.example.com')
+    t.equal(appConfig.scopes.length, 1)
+    t.equal(appConfig.scopes[0], 'store_write')
+    t.equal(appConfig.manifestURI(), 'https://custom.example.com/manifest.json')
+    t.equal(appConfig.redirectURI(), 'https://custom.example.com')
   })
 }